app/requirements.txt
nicegui==1.4.29
loguru~=0.7.2
google-cloud-storage==2.17.0
celery[redis]==5.4.0
pydantic==2.8.2